requirements.txt
django-ninja == 1.1.0
django == 5.0.4
pillow ==10.4.0